Revert "Move isVSCodeExtension func to avoid circular dependency (#7938)"
This reverts commit e7d679dc9af55627fbc178ee705dcc000d152757.
