--- src/vsCodeUtils.ts
@@ -1,8 +1,0 @@
-import { env } from "process";
-
-/**
- * Detect if code is running in a VSCode Extension
- */
-export function isVSCodeExtension(): boolean {
-  return !!env.VSCODE_CWD;
-}
